+++ blockapps-rest/ReleaseNotes.md
@@ -5,6 +5,10 @@
 - Add source mappings for minified dist code
 - Add babel-preset-minify, remove gulp-uglify
 
+#### Minor upgrades
+
+- Added Bloc /compile endpoint
+
 ### Version 7.6.3
 
 - Fix for util.uid() to always return string value for compatibility with STRATO 5.2+